+++ app/core/Wallet/AccountList/style.css
