+++ app/core/Wallet/History/style.css
