+++ app/core/Wallet/Receive/style.css
