+++ app/core/Wallet/images/icnSettings.png
