+++ app/core/Wallet/style.css
