+++ app/core/Wallet/styledComponents.js
