+++ app/routes.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { Switch } from 'react-router-dom'
 import Route from './common/auth/route'
 import Launch from './core/Launch'
-import Main from './core/Main'
+import Wallet from './core/Wallet'
 import AppView from './core/AppView'
 import Creation from './core/Creation'
 import Restore from './core/Restore'
